Made Some Last Minute Changes After Some Feedback, 4/9/2024
Made Some Last Minute Changes After Some Feedback. The biggest one being adding an entire new set of sound-clips, that are just a lower-quality more "crusty" sounding version of the original sounds, these are essentially the "retro" sounds, including an option to swap between these at anytime. Also made it so you could set the volume settings to 10x at max rather than only 5x, few other small changes as well,. But NOW this will be the version I release onto the Nexus and to the public when I get the chance to, 4/9/2024.