v1.01 Quick Patch So Exceptions Are Better Handled, 4/13/2024
v1.01 Quick Patch So Exceptions Are Better Handled. It seems the mod 'tempered interiors' was causing this mod to throw and exception when transitioning into interiors. So I just added some extra exception handling code to hopefully prevent cascading errors elsewhere when/if this occurs. Not a perfect "fix", but at the very least should prevent loading issues and other things, even if the footsteps are not accurate when these exceptions occur, rather that than a save not loading properly, etc, 4/13/2024.